Actually, NoChainTip is used in production for isolated connections
